# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: customers
    description: Customer operations
  - name: subscriptions
    description: Subscription operations
  - name: invoices
    description: Invoice operations
  - name: dashboard
    description: Dashboard operations
  - name: billing
    description: Billing operations
  - name: webhooks
    description: Webhook operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Customer not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Customer deleted

  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List all subscriptions
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, cancelled, overdue]
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create a new subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionBody"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  /subscriptions/{id}:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get subscription by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Update a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionBody"
      responses:
        "200":
          description: Subscription updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    delete:
      operationId: deleteSubscription
      tags: [subscriptions]
      summary: Cancel a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Subscription cancelled

  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: subscriptionId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, overdue, cancelled]
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"

  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  /invoices/{id}/generate-pix:
    post:
      operationId: generatePixCharge
      tags: [billing]
      summary: Generate PIX charge for invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PIX charge generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PixCharge"

  /invoices/{id}/send-reminder:
    post:
      operationId: sendPaymentReminder
      tags: [billing]
      summary: Send payment reminder via WhatsApp
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reminder sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary metrics
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/recent-payments:
    get:
      operationId: getRecentPayments
      tags: [dashboard]
      summary: Get recent payments
      responses:
        "200":
          description: Recent payments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecentPayment"

  /dashboard/monthly-revenue:
    get:
      operationId: getMonthlyRevenue
      tags: [dashboard]
      summary: Get monthly revenue data for charts
      responses:
        "200":
          description: Monthly revenue data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyRevenue"

  /billing/process-due:
    post:
      operationId: processDueSubscriptions
      tags: [billing]
      summary: Process subscriptions due today (cron job)
      responses:
        "200":
          description: Processing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CronResult"

  /webhooks/qqpag:
    post:
      operationId: qqpagWebhook
      tags: [webhooks]
      summary: QQPag payment confirmation webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QQPagWebhookPayload"
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        whatsapp:
          type: string
        cpfCnpj:
          type: string
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - whatsapp
        - cpfCnpj
        - status
        - createdAt
        - updatedAt

    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        whatsapp:
          type: string
        cpfCnpj:
          type: string
        status:
          type: string
          enum: [active, inactive]
      required:
        - name
        - whatsapp
        - cpfCnpj

    UpdateCustomerBody:
      type: object
      properties:
        name:
          type: string
        whatsapp:
          type: string
        cpfCnpj:
          type: string
        status:
          type: string
          enum: [active, inactive]

    Subscription:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        plan:
          type: string
        periodicity:
          type: string
          enum: [monthly, annual]
        amount:
          type: number
        status:
          type: string
          enum: [active, cancelled, overdue]
        nextBillingDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - plan
        - periodicity
        - amount
        - status
        - nextBillingDate
        - createdAt
        - updatedAt

    CreateSubscriptionBody:
      type: object
      properties:
        customerId:
          type: integer
        plan:
          type: string
        periodicity:
          type: string
          enum: [monthly, annual]
        amount:
          type: number
        nextBillingDate:
          type: string
          format: date-time
      required:
        - customerId
        - plan
        - periodicity
        - amount
        - nextBillingDate

    UpdateSubscriptionBody:
      type: object
      properties:
        plan:
          type: string
        periodicity:
          type: string
          enum: [monthly, annual]
        amount:
          type: number
        status:
          type: string
          enum: [active, cancelled, overdue]
        nextBillingDate:
          type: string
          format: date-time

    Invoice:
      type: object
      properties:
        id:
          type: integer
        subscriptionId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, overdue, cancelled]
        dueDate:
          type: string
          format: date-time
        paidAt:
          type: ["string", "null"]
          format: date-time
        pixCode:
          type: ["string", "null"]
        pixQrCode:
          type: ["string", "null"]
        externalId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - subscriptionId
        - customerId
        - amount
        - status
        - dueDate
        - createdAt
        - updatedAt

    PixCharge:
      type: object
      properties:
        invoiceId:
          type: integer
        qrCode:
          type: string
        pixCopiaECola:
          type: string
        externalId:
          type: string
      required:
        - invoiceId
        - qrCode
        - pixCopiaECola
        - externalId

    MessageResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    DashboardSummary:
      type: object
      properties:
        mrr:
          type: number
        activeSubscriptions:
          type: integer
        churnRate:
          type: number
        todayRevenue:
          type: number
        pendingInvoices:
          type: integer
        overdueInvoices:
          type: integer
      required:
        - mrr
        - activeSubscriptions
        - churnRate
        - todayRevenue
        - pendingInvoices
        - overdueInvoices

    RecentPayment:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        amount:
          type: number
        paidAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - amount
        - paidAt

    MonthlyRevenue:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
      required:
        - month
        - revenue

    CronResult:
      type: object
      properties:
        processed:
          type: integer
        successful:
          type: integer
        failed:
          type: integer
        message:
          type: string
      required:
        - processed
        - successful
        - failed
        - message

    QQPagWebhookPayload:
      type: object
      properties:
        transactionId:
          type: string
        externalId:
          type: string
        status:
          type: string
        amount:
          type: number
        paidAt:
          type: string
      required:
        - transactionId
        - externalId
        - status

    WebhookResult:
      type: object
      properties:
        received:
          type: boolean
      required:
        - received