# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  mrr: number;
  activeSubscriptions: number;
  churnRate: number;
  todayRevenue: number;
  pendingInvoices: number;
  overdueInvoices: number;
}