# lib/api-zod/src/generated/types/recentPayment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecentPayment {
  id: number;
  customerName: string;
  amount: number;
  paidAt: Date;
}